+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log for SD.Next
 
+## Update for 2024-06-14
+
+- force apply vae config on model load
+
 ## Update for 2024-06-13
 
 ### Highlights for 2024-06-13